refactor(database): move setup-test-db to src/testing for shared consumption
