+++ core/database/__tests__/repositories/usage-events-repository.test.ts
@@ -1,6 +1,6 @@
 import { beforeAll, afterAll, describe, it, expect } from "vitest";
 import { seed } from "drizzle-seed";
-import { setupTestDb } from "../helpers/setup-test-db";
+import { setupTestDb } from "../../src/testing/setup-test-db";
 import * as schema from "@core/database/schema";
 import { upsertUsageEventSchema } from "@core/database/schemas/usage-events";
 import { meters } from "@core/database/schemas/meters";
